refactor: add interface documentation and change experationTime type

+++ src/domain/usecases/tokenizer_service.go
@@ -3,5 +3,7 @@ package usecases
 import "github.com/google/uuid"
 
 type TokenizerService interface {
-	Token(userID uuid.UUID, expirationAccessTime int) (string, error)
+	// Token generates a new JWT token with userID and experionTime properties.
+	// The expirationAccessTime has seconds unit of time.
+	Token(userID uuid.UUID, expirationAccessTime float64) (string, error)
 }
